added an rdf a filter for collections
