+++ lib/jekyll/filters/rdf_property.rb
@@ -29,6 +29,7 @@ module Jekyll
   # Internal module to hold the medthod #rdf_property
   #
   module RdfProperty
+    include Jekyll::RdfPrefixResolver
     ##
     # Computes all objects for which statements exist containing the given subject and predicate and returns any of them
     #
@@ -81,26 +82,6 @@ module Jekyll
         return (s.term.language == lang.to_sym)
       else
         return false
-      end
-    end
-
-    private
-    def rdf_resolve_prefix(input, predicate)
-      if(predicate[0] == "<" && predicate[-1] == ">")
-        return predicate[1..-2]
-      end
-      arr=predicate.split(":",2)  #bad regex, would also devide 'http://example....' into 'http' and '//example....',even though it is already a complete URI; if 'PREFIX http: <http://...> is defined, 'http' in 'http://example....' could be mistaken for a prefix
-      if((arr[1].include? (":")) || (arr[1][0..1].eql?("//")))
-        raise UnMarkedUri.new(predicate, input.page.data['template'])
-      end
-      if(!input.page.data["rdf_prefixes"].nil?)
-        if(!input.page.data["rdf_prefix_map"][arr[0]].nil?)
-          return arr[1].prepend(input.page.data["rdf_prefix_map"][arr[0]])
-        else
-          raise NoPrefixMapped.new(predicate, input.page.data['template'], arr[0])
-        end
-      else
-        raise NoPrefixesDefined.new(predicate, input.page.data['template'])
       end
     end
   end
