+++ test/source/_layouts/abraham.html
@@ -1,4 +1,5 @@
 ---
+rdf_prefix_path: simpsons.pref
 layout: abraham_outer_layout
 ---
