--- test/source/index.html
@@ -1,33 +1,0 @@
----
-layout: default
----
-
-<div class="home">
-
-  <h1 class="page-heading">Posts</h1>
-
-  <ul class="post-list">
-    {% for post in site.posts %}
-      <li>
-        <span class="post-meta">{{ post.date | date: "%b %-d, %Y" }}</span>
-
-        <h2>
-          <a class="post-link" href="{{ post.url | prepend: site.baseurl }}">{{ post.title }}</a>
-        </h2>
-      </li>
-    {% endfor %}
-  </ul>
-
-  <table>
-    {% for statement in site.data.rdf %}
-      <tr>
-        <td>{{ statement[0] }}</td> <!-- Subject -->
-        <td>{{ statement[1] }}</td> <!-- Predicate -->
-        <td>{{ statement[2] }}</td> <!-- Object -->
-      </tr>
-    {% endfor %}
-  </table>
-
-  <p class="rss-subscribe">subscribe <a href="{{ "/feed.xml" | prepend: site.baseurl }}">via RSS</a></p>
-
-</div>
